+++ components/forms/UploadForm.vue
@@ -31,21 +31,18 @@
           class="border p-4 rounded"
         >
           <div class="font-bold">{{ formatDate(date) }}</div>
-          <div
-            v-for="(timeSlot, index) in timeSlots"
-            :key="index"
-            class="text-blue-500"
-          >
-            {{ timeSlot.start }} ~ {{ timeSlot.end }}
+          <div class="text-blue-500 whitespace-pre-line">
+            {{ formatTimeForDisplay(timeSlots) }}
           </div>
         </div>
       </div>
 
       <div class="mt-6 flex justify-end gap-4 sticky bottom-0 bg-white pb-6">
         <buttons-square
-          @click="copyToClipboard"
+          @click="handleCopy"
           label="コピー"
           color="bg-gray-300"
+          :isUse="Object.keys(displayData).length > 0"
         />
         <buttons-square
           @click="syncData"
@@ -61,6 +58,8 @@
 
 <script setup>
 import { onMounted, onUnmounted, ref, watch } from "vue";
+import { useTimeUtils } from "@/utils/TimeUtils";
+import { copyToClipboard } from "@/utils/CopyDate";
 
 const props = defineProps({
   timeData: {
@@ -71,6 +70,7 @@ const props = defineProps({
 
 const emit = defineEmits(["close"]);
 const displayData = ref({});
+const { formatTimeForDisplay } = useTimeUtils();
 
 watch(
   () => props.timeData,
@@ -86,33 +86,8 @@ const handleEscKey = (event) => {
   }
 };
 
-const formatDate = (dateString) => {
-  const date = new Date(dateString);
-  const weekdays = ["日", "月", "火", "水", "木", "金", "土"];
-  return `${date.getFullYear()}年${date.getMonth() + 1}月${date.getDate()}日(${
-    weekdays[date.getDay()]
-  })`;
-};
-
-const copyToClipboard = () => {
-  if (Object.keys(displayData.value).length === 0) {
-    alert("コピーするデータがありません");
-    return;
-  }
-
-  const text = Object.entries(displayData.value)
-    .map(([date, timeSlots]) => {
-      const timeStrings = timeSlots
-        .map((time) => `${time.start} ~ ${time.end}`)
-        .join(",");
-      return `${formatDate(date)}:  ${timeStrings}`;
-    })
-    .join("\n");
-
-  navigator.clipboard
-    .writeText(text)
-    .then(() => alert("クリップボードにコピーしました"))
-    .catch((err) => console.error("コピーに失敗しました:", err));
+const handleCopy = () => {
+  copyToClipboard(displayData.value);
 };
 
 const syncData = async () => {
